User/create

+++ app/Classes/FileHelper.php
@@ -25,10 +25,9 @@ class FileHelper
      * OSave $file to $path directory
      * @param UploadedFile $file
      * @param string $path
-     * @param string|null $settings_key
      * @return string
      */
-    public static function saveFile(UploadedFile $file, string $path, ?string $settings_key = null): string
+    public static function saveFile(UploadedFile $file, string $path): string
     {
         $path = FileHelper::createFolder(public_path($path));
         $filename = $file->getClientOriginalName();
@@ -65,22 +64,24 @@ class FileHelper
      */
     public static function recursiveRemove(string $path): void
     {
-        if (is_file($path)) {
-            unlink($path);
-        } else {
-            $files = new DirectoryIterator($path);
+        if (file_exists($path)) {
+            if (is_file($path)) {
+                unlink($path);
+            } else {
+                $files = new DirectoryIterator($path);
 
-            foreach ($files as $file) {
-                if (!$file->isDot()) {
-                    if ($file->isDir()) {
-                        self::recursiveRemove($file->getPathname());
-                    } else {
-                        unlink($file->getPathname());
+                foreach ($files as $file) {
+                    if (!$file->isDot()) {
+                        if ($file->isDir()) {
+                            self::recursiveRemove($file->getPathname());
+                        } else {
+                            unlink($file->getPathname());
+                        }
                     }
                 }
-            }
 
-            rmdir($path);
+                rmdir($path);
+            }
         }
     }
 }
